requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-ipware==7.0.1
django-rest-authemail==2.1.7
django-timezone-field==7.0
djangorestframework==3.15.2
google-auth==2.37.0
idna==3.10
kombu==5.4.2
mysqlclient==2.2.6
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-ipware==3.0.0
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
